+++ crest/Assets/Crest/Crest/Scripts/Spline/Spline.cs
@@ -26,8 +26,17 @@ namespace Crest.Spline
         /// </summary>
         [SerializeField, HideInInspector]
 #pragma warning disable 414
-        int _version = 0;
+        int _version = 1;
 #pragma warning restore 414
+
+        public enum Offset
+        {
+            Left,
+            Center,
+            Right
+        }
+        [Tooltip("Where generated ribbon should lie relative to spline. If set to Center, ribbon is centered around spline.")]
+        public Offset _offset = Offset.Center;
 
         [Tooltip("Connect start and end point to close spline into a loop. Requires at least 3 spline points.")]
         public bool _closed = false;
@@ -36,12 +45,9 @@ namespace Crest.Spline
         float _radius = 20f;
         [SerializeField, Delayed]
         int _subdivisions = 1;
-        [SerializeField, Delayed]
-        int _smoothingIterations = 0;
 
         public float Radius => _radius;
         public int Subdivisions => _subdivisions;
-        public int SmoothingIterations => _smoothingIterations;
 
         public bool AttachDataToSplinePoint(GameObject splinePoint)
         {
@@ -53,6 +59,28 @@ namespace Crest.Spline
 
             splinePoint.AddComponent<SplinePointData>();
             return true;
+        }
+    }
+
+    // Version handling - perform data migration after data loaded.
+    public partial class Spline : ISerializationCallbackReceiver
+    {
+        public void OnBeforeSerialize()
+        {
+            // Intentionally left empty.
+        }
+
+        public void OnAfterDeserialize()
+        {
+            // Version 1 (2021.10.02)
+            // - Alignment added with default value different than old behaviour
+            if (_version == 0)
+            {
+                // Set alignment to Right to maintain old behaviour
+                _offset = Offset.Right;
+
+                _version = 1;
+            }
         }
     }